+++ settings/fonts.php
@@ -41,11 +41,19 @@ if ($ADMIN->fulltree) {
     $setting = new admin_setting_heading($name, $heading, '');
     $page->add($setting);
 
-    // Main Google Font Name.
+    // Google fonts.
+    $name = 'theme_adaptable/googlefonts';
+    $title = get_string('googlefonts', 'theme_adaptable');
+    $description = get_string('googlefontsdesc', 'theme_adaptable');
+    $default = false;
+    $setting = new admin_setting_configcheckbox($name, $title, $description, $default, true, false);
+    $page->add($setting);
+
+    // Main Font Name.
     $name = 'theme_adaptable/fontname';
     $title = get_string('fontname', 'theme_adaptable');
     $description = get_string('fontnamedesc', 'theme_adaptable');
-    $default = 'Lexend Deca';
+    $default = 'sans-serif';
     $setting = new admin_setting_configselect($name, $title, $description, $default, $fontlist);
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
@@ -123,7 +131,7 @@ if ($ADMIN->fulltree) {
     $name = 'theme_adaptable/fontheadername';
     $title = get_string('fontheadername', 'theme_adaptable');
     $description = get_string('fontheadernamedesc', 'theme_adaptable');
-    $default = 'Archivo';
+    $default = 'sans-serif';
     $setting = new admin_setting_configselect($name, $title, $description, $default, $fontlist);
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
@@ -148,7 +156,7 @@ if ($ADMIN->fulltree) {
     $name = 'theme_adaptable/fonttitlename';
     $title = get_string('fonttitlename', 'theme_adaptable');
     $description = get_string('fonttitlenamedesc', 'theme_adaptable');
-    $default = 'Cabin';
+    $default = 'sans-serif';
     $setting = new admin_setting_configselect($name, $title, $description, $default, $fontlist);
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);